+++ include/ui/user_interface.hpp
@@ -84,7 +84,7 @@ public:
   /// \param _s current score.
   /// \param _mouse current mouse position.
   //----------------------------------------------------------------------------------------------------------------------
-  void update(int _s, const vec2 _mouse);
+  void update(float _s, const vec2 _mouse);
 
   //----------------------------------------------------------------------------------------------------------------------
   /// \brief Clears the UI.
