+++ src/ui/user_interface.cpp
@@ -50,7 +50,7 @@ void userInterface::reset()
     for(auto &i : m_elements) i.reset();
 }
 
-void userInterface::update(const int _s, const vec2 _mouse)
+void userInterface::update(const float _s, const vec2 _mouse)
 {
     for(auto &i : m_elements) i.update(_s, _mouse);
 }
